refactor metrics suites
